Extract WASM section model
